backend/requirements.txt
django
django-cors-headers
djangorestframework
markdown
django-filter
openai 
numpy